feat: #2943 developer about section to main, contact section
