+++ packages/developer-portal/src/components/pages/app-detail-preview/client/aside.tsx
@@ -1,14 +1,13 @@
 import React from 'react'
 import { DesktopIntegrationTypeModel, AppDetailModel } from '@reapit/foundations-ts-definitions'
 import { FlexContainerBasic } from '@reapit/elements'
-import { ContactDeveloperSection } from './contact-developer-modal'
+import { ContactDeveloperSection } from './contact-developer-section'
 import useReactResponsive from '@/components/hooks/use-react-responsive'
 import { History } from 'history'
 import routes from '@/constants/routes'
 import { CategorySection, DesktopIntegrationSection, DirectApiSection, BackToAppsSection } from '../common/ui-sections'
 
 import { useHistory } from 'react-router'
-import { DeveloperAboutSection } from '../../app-detail/app-sections'
 
 interface AsideProps {
   appDetailData: AppDetailModel
@@ -22,14 +21,13 @@ export const onBackToAppsButtonClick = (history: History) => {
 }
 
 export const Aside: React.FC<AsideProps> = ({ desktopIntegrationTypes, appDetailData }) => {
-  const { category, developer, telephone, supportEmail, homePage, isDirectApi, developerAbout } = appDetailData
+  const { category, developer, telephone, supportEmail, homePage, isDirectApi } = appDetailData
 
   const { isMobile } = useReactResponsive()
   const history = useHistory()
 
   return (
     <FlexContainerBasic flexColumn hasPadding hasBackground isFullHeight={!isMobile}>
-      {developerAbout && <DeveloperAboutSection isSidebar>{developerAbout}</DeveloperAboutSection>}
       <CategorySection category={category} isSidebar />
       <DesktopIntegrationSection desktopIntegrationTypes={desktopIntegrationTypes} isSidebar />
       <DirectApiSection isDirectApi={isDirectApi} isSidebar />
