+++ packages/marketplace/src/components/pages/app-detail/client/__tests__/contact-developer-section.test.tsx
@@ -1,10 +1,5 @@
 import React from 'react'
-import {
-  ContactDeveloperSection,
-  ContactDeveloperSectionType,
-  openContactModal,
-  closeContactModal,
-} from '../contact-developer-modal'
+import { ContactDeveloperSection, ContactDeveloperSectionType } from '../contact-developer-section'
 import { shallow } from 'enzyme'
 
 const props: ContactDeveloperSectionType = {
@@ -23,19 +18,5 @@ describe('ContactDeveloperSection', () => {
 
   test('should match snapshot with gutter', () => {
     expect(shallow(<ContactDeveloperSection {...props} hasGutter />)).toMatchSnapshot()
-  })
-
-  test('handle openContactModal', () => {
-    const setVisible = jest.fn()
-    openContactModal(setVisible)()
-    expect(setVisible).toBeCalled()
-    expect(setVisible).toBeCalledWith(true)
-  })
-
-  test('handle closeContactModal', () => {
-    const setVisible = jest.fn()
-    closeContactModal(setVisible)()
-    expect(setVisible).toBeCalled()
-    expect(setVisible).toBeCalledWith(false)
   })
 })
