+++ packages/marketplace/src/components/pages/app-detail/client/app-content.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react'
 import { Section } from '@reapit/elements'
-import { SummarySection, AdditionalImagesSection, PermissionsSection, DescriptionSection } from '../common/ui-sections'
+import {
+  SummarySection,
+  AdditionalImagesSection,
+  PermissionsSection,
+  DescriptionSection,
+  DeveloperAboutSection,
+} from '../common/ui-sections'
 import { AppDetailModel } from '@reapit/foundations-ts-definitions'
 
 export type AppContentProps = {
@@ -8,7 +14,7 @@ export type AppContentProps = {
 }
 
 const AppContent: React.FC<AppContentProps> = ({ appDetailData }) => {
-  const { summary = '', description = '', scopes = [], media = [] } = appDetailData
+  const { summary = '', description = '', scopes = [], media = [], developerAbout } = appDetailData
 
   return (
     <Section isFlex isFlexColumn hasPadding={false} hasMargin={false}>
@@ -16,6 +22,7 @@ const AppContent: React.FC<AppContentProps> = ({ appDetailData }) => {
       <AdditionalImagesSection images={media} splitIndex={1} numberImages={2} />
       <DescriptionSection description={description} />
       <AdditionalImagesSection images={media} splitIndex={3} numberImages={2} />
+      {developerAbout && <DeveloperAboutSection>{developerAbout}</DeveloperAboutSection>}
       <PermissionsSection permissions={scopes} />
     </Section>
   )
