+++ packages/marketplace/src/components/pages/app-detail/client/aside.tsx
@@ -1,18 +1,12 @@
 import React from 'react'
 import { DesktopIntegrationTypeModel, AppDetailModel } from '@reapit/foundations-ts-definitions'
 import { FlexContainerBasic } from '@reapit/elements'
-import { ContactDeveloperSection } from './contact-developer-modal'
+import { ContactDeveloperSection } from './contact-developer-section'
 import useReactResponsive from '@/components/hooks/use-react-responsive'
 import { History } from 'history'
 import routes from '@/constants/routes'
 // Commenting out for now until we have the API
-import {
-  DeveloperSection /*, DeveloperAboutSection */,
-  CategorySection,
-  DesktopIntegrationSection,
-  DirectApiSection,
-  BackToAppsSection,
-} from '../common/ui-sections'
+import { CategorySection, DesktopIntegrationSection, DirectApiSection, BackToAppsSection } from '../common/ui-sections'
 import { useHistory } from 'react-router'
 import WebComponentConfig from './web-component-config-modal'
 
@@ -34,7 +28,6 @@ export const Aside: React.FC<AsideProps> = ({ desktopIntegrationTypes, appDetail
   const history = useHistory()
   return (
     <FlexContainerBasic flexColumn hasPadding hasBackground isFullHeight={!isMobile}>
-      {developer && <DeveloperSection developer={developer} isSidebar />}
       <CategorySection category={category} isSidebar />
       <DesktopIntegrationSection desktopIntegrationTypes={desktopIntegrationTypes} isSidebar />
       <DirectApiSection isDirectApi={isDirectApi} isSidebar />
