+++ packages/marketplace/src/components/pages/app-detail/common/ui-sections.tsx
@@ -63,10 +63,6 @@ interface InstallationsTableSectionProps extends IsSidebar {
   columns: any[]
 }
 
-interface DeveloperSectionProps extends IsSidebar {
-  developer: string
-}
-
 interface AdditionalImagesSectionProps {
   images: MediaModel[]
   splitIndex: number
@@ -175,9 +171,9 @@ export const InstallationsTableSection: React.FC<InstallationsTableSectionProps>
   )
 }
 
-export const DeveloperSection: React.FC<DeveloperSectionProps> = ({ isSidebar = false, developer }) => (
-  <AppDetailSection headerText="Developer Links" isSidebar={isSidebar}>
-    {developer}
+export const DeveloperAboutSection: React.FC<IsSidebar> = ({ isSidebar = false, children }) => (
+  <AppDetailSection headerText="About Developer" isSidebar={isSidebar}>
+    {children}
   </AppDetailSection>
 )
